Updates user profile picture function
Refactors the Set-VerkadaAccessUserProfilePicture function to use the public API.

This change updates the function to utilize the public API for setting user profile pictures,
enhancing functionality and security. It replaces the internal API call with a call to
https://apidocs.verkada.com/reference/putprofilephotoviewv1, adding support for externalId.
It also adds support for selecting which region of the public API to use.