Adds Get-VerkadaAccessEvents function
Adds a new function to retrieve Verkada Access Control events.

This function allows users to query access events within a specified time range and for specific event types, sites, devices, or users. It leverages the Verkada API to fetch the events and returns them as PowerShell objects. The function includes error handling and the option to write errors to a file.

Also updates the module manifest to include the new function and its alias.